backend/requirements.txt
# Backend dependencies for LLM Video Analytics
# Python version: 3.10+
# All packages are pinned to specific versions for reproducibility

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
hypercorn[h2]>=0.17.0
python-multipart==0.0.6
slowapi==0.1.9  # Rate limiting

# Data Validation & Settings
pydantic>=2.9.0
pydantic-settings>=2.0.0

# LLM Integration
google-genai>=1.0.0
google-auth>=2.0.0
langchain-google-genai>=4.2.0

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
python-dotenv==1.0.1

# HTTP Client
httpx>=0.27.0

# Logging & Monitoring
# python-json-logger==2.0.7

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black==23.12.1  # Code formatter
flake8==6.1.0  # Linter
isort==5.13.2  # Import sorter
